pyproject.toml
[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[project]
name = "tw-odc"
version = "0.1.0"
description = "Taiwan Open Data Checker — 台灣開放資料品質檢測工具"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "chardet>=7.0.1",
    "i18nice>=0.16.0",
    "jsonpatch>=1.33",
    "platformdirs>=4.0",
    "python-magic>=0.4.27",
    "rich>=13.0.0",
    "typer>=0.24.1",
]

[project.scripts]
tw-odc = "tw_odc.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
